+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Data is not stored using the repository id as key ([#12](https://github.com/scm-manager/scm-statistic-plugin/pull/12))
+
 ## 2.1.0 - 2020-12-17
 ### Added
 - Mark read only verbs to be able to see statistics in archived repositories ([#10](https://github.com/scm-manager/scm-statistic-plugin/pull/10))